+++ byte.h
@@ -1,9 +1,9 @@
 #include <stdint.h>
 
-void poly2bytes(unsigned char *b, const int16_t * f);
+void poly2bytes(unsigned char *b, const int64_t * f);
 
-void bytes2poly(int16_t *f, const unsigned char *b);
+void bytes2poly(int64_t *f, const unsigned char *b);
 
 void displaybits(unsigned char *b);
 
-void display10bits(int16_t *f);
+void display10bits(int64_t *f);
